% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_combine.R
\name{qm_combine}
\alias{qm_combine}
\title{Combine objects}
\usage{
qm_combine(...)
}
\arguments{
\item{...}{A list of cluster objects to be combined.}
}
\value{
A single tibble with all observations from the listed cluster objects.
}
\description{
A wrapper around \code{dplyr::bind_rows} for combining cluster objects into a single tibble.
}
