% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_create.R
\name{qm_create}
\alias{qm_create}
\title{Create cluster object}
\usage{
qm_create(ref, key, value, rid, cid, category, ...)
}
\arguments{
\item{ref}{An sf object that serves as a master list of features}

\item{key}{Name of geographic id variable in the \code{ref} object to match input values to}

\item{value}{A vector of input values}

\item{rid}{Respondent identification number; a user defined integer value that uniquely identifies respondents
in the project}

\item{cid}{Cluster identification number; a user defined interger value that unqiuely identifies clusters}

\item{category}{Category type; a user defined value that describes what the cluster represents}

\item{...}{An unquoted list of variables from the sf object to include in the out}
}
\value{
A tibble with the cluster values merged with elements of the reference data.
}
\description{
Each vector of input values is converted to a tibble organized in a "tidy" fashion.
}
\details{
A cluster object contains a row for each feature in the reference data set. The \code{key} variable
values are included in a variable named identically to the \code{key}. Additionally, a variable called
\code{COUNT} is created. This will always be equal to \code{1} for each observation. It is intended to be
used for summarization later in the data analysis process. Finally, three pieces of metadata are also included
as arguments to provide data for subsetting later: a respondent identification number (\code{rid}), a
cluster identification number (\code{cid}), and a category for the cluster type (\code{category}). These
arguments are converted into values for the output variables \code{RID}, \code{CID}, and \code{CAT} respectively.
}
