% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_preview.R
\name{qm_preview}
\alias{qm_preview}
\title{Preview Input}
\usage{
qm_preview(.data, key, value)
}
\arguments{
\item{.data}{An sf object}

\item{key}{Quoted name of id variable to match input values to}

\item{value}{A vector of input values}
}
\value{
An interactive leaflet map with the values from the input vector highlighted in red.
}
\description{
This function renders the input vector as a polygon shapefile using the leaflet package.
}
