% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_summarize.R
\name{qm_summarize}
\alias{qm_summarize}
\title{Summarize Clusters}
\usage{
qm_summarize(clusters, key, category, ref)
}
\arguments{
\item{clusters}{A tibble created by \code{qm_create} with two or more clusters worth of data}

\item{key}{Name of geographic id variable in the tibble specified in \code{clusters}}

\item{category}{Value of the \code{CAT} variable to be analyzed}

\item{ref}{Optional sf object; a join will be performed between the sf object and the summarized clusters}
}
\value{
A tibble or a sf object (if \code{ref} is specified) that contains a count of the number of clusters
a given feature is included in.
}
\description{
This function creates a column that contains a single observation for each unique value
in the key variable. For each feature, a count corresponding to the number of times that feature is
identified in a cluster for the give category is also provided.
}
\seealso{
qm_create
}
