% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_validate.R
\name{qm_validate}
\alias{qm_validate}
\title{Validate input vector}
\usage{
qm_validate(ref, key, value)
}
\arguments{
\item{ref}{An sf object that serves as a master list of features}

\item{key}{Quoted name of id variable in the \code{ref} object to match input values to}

\item{value}{A vector of input values}
}
\value{
A logical scalar that is \code{TRUE} is all input values match values in the key variable.
}
\description{
This function ensures that the input vector values match valid values in a source shapefile.
}
