% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_validate_obj.R
\name{qm_validate_obj}
\alias{qm_validate_obj}
\title{Validate cluster object}
\usage{
qm_validate_obj(obj)
}
\arguments{
\item{obj}{Object to test}
}
\value{
A logical scalar that is \code{TRUE} is the given object is of class \code{qm_cluster};
it will return \code{FALSE} otherwise.
}
\description{
This function tests to see whether an object is of class \code{qm_cluster}.
}
