% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_verify.R
\name{qm_verify}
\alias{qm_verify}
\title{Verify Previously Saved Cluster Data}
\usage{
qm_verify(clusters)
}
\arguments{
\item{clusters}{An object created by \code{qm_combine} with two or more clusters worth of data
that has been previously saved and requires verification before summarization.}
}
\value{
A tibble stored with a custom class of \code{qm_cluster} to facilitate data validation.
}
\description{
Users may wish to save long-form combined cluster data as a \code{.csv} file
    or similar after combining individual clusters with \code{\link{qm_combine}}. Unless
    data are saved as an \code{.rda} file, the class definition of \code{"qm_cluster"}
    will be lost. The \code{qm_verify} function allows users to import data from any
    file type readable by \code{R}, verify that it has the column names needed for
    \code{\link{qm_summarize}}, and re-add the \code{"qm_cluster"} class.
}
