% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualmap-package.R
\docType{package}
\name{qualmap-package}
\alias{qualmap}
\alias{qualmap-package}
\title{qualmap: Opinionated Approach for Digitizing Semi-Structured Qualitative GIS Data}
\description{
Provides a set of functions for taking qualitative GIS data, hand drawn on a map, and converting it to a simple features object. These tools are focused on data that are drawn on a map that contains some type of polygon features. For each area identified on the map, the id numbers of these polygons can be entered as vectors and transformed using qualmap.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://chris-prener.github.io/qualmap/}
  \item Report bugs at \url{https://github.com/chris-prener/qualmap/issues}
}

}
\author{
\strong{Maintainer}: Christopher Prener \email{chris.prener@gmail.com} (\href{https://orcid.org/0000-0002-4310-9888}{ORCID})

}
\keyword{internal}
