% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stLouis.R
\docType{data}
\name{stLouis}
\alias{stLouis}
\title{St. Louis Census Tracts, 2016}
\format{A data frame with 106 rows and 7 variables:
\describe{
  \item{STATEFP}{state FIPS code}
  \item{COUNTYFP}{county FIPS code}
  \item{TRACTCE}{tract FIPS code}
  \item{GEOID}{full GEOID string}
  \item{NAME}{tract FIPS code, decimal}
  \item{NAMELSAD}{tract name}
  \item{geometry}{simple features geometry}
  }}
\source{
\href{https://www.census.gov}{U.S. Census Bureau}

#' @examples
str(stLouis)
head(stLouis)
}
\usage{
data(stLouis)
}
\description{
A simple features data set containing the geometry and associated attributes for the 2016 City of St. Louis
census tracts.
}
\note{
These data have been modified from the full version available from the Census Bureau - some
variables related to geometry and geography type have been removed.
}
\keyword{datasets}
