% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_preview.R
\name{qm_preview}
\alias{qm_preview}
\title{Preview Input}
\usage{
qm_preview(ref, key, value)
}
\arguments{
\item{ref}{An sf object that serves as a master list of features}

\item{key}{Name of geographic id variable in the \code{ref} object to match input values to}

\item{value}{A vector of input values}
}
\value{
An interactive leaflet map with the values from the input vector highlighted in red.
}
\description{
This function renders the input vector as a polygon shapefile using the leaflet package.
}
\examples{
\dontrun{
stl <- stLouis
stl <- mutate(stl, TRACTCE = as.numeric(TRACTCE))

cluster <- qm_define(118600, 119101, 119300)
qm_preview(ref = stl, key = TRACTCE, value = cluster)
}

}
