% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_summarize.R
\name{qm_summarize}
\alias{qm_summarize}
\title{Summarize Clusters}
\usage{
qm_summarize(ref, key, clusters, category)
}
\arguments{
\item{ref}{An \code{sf} object that serves as a master list of features}

\item{key}{Name of geographic id variable in the \code{ref} object to match input values to}

\item{clusters}{A tibble created by \code{qm_combine} with two or more clusters worth of data}

\item{category}{Value of the \code{CAT} variable to be analyzed}
}
\value{
A tibble or a \code{sf} object (if geometry = \code{TRUE}) that contains a count of the number
of clusters a given feature is included in.
}
\description{
This function creates a column that contains a single observation for each unique value
in the key variable. For each feature, a count corresponding to the number of times that feature is
identified in a cluster for the give category is also provided.
}
\examples{
# load and format reference data
stl <- stLouis
stl <- dplyr::mutate(stl, TRACTCE = as.numeric(TRACTCE))

# create clusters
cluster1 <- qm_define(118600, 119101, 119300)
cluster2 <- qm_define(119300, 121200, 121100)

# create cluster objects
cluster_obj1 <- qm_create(ref = stl, key = TRACTCE, value = cluster1,
    rid = 1, cid = 1, category = "positive")
cluster_obj2 <- qm_create(ref = stl, key = TRACTCE, value = cluster2,
    rid = 1, cid = 2, category = "positive")

# combine cluster objects
clusters <- qm_combine(cluster_obj1, cluster_obj2)

# summarize cluster objects
positive <- qm_summarize(ref = stl, key = TRACTCE, clusters = clusters, category = "positive")

}
\seealso{
\code{qm_combine}
}
