% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qm_validate.R
\name{qm_validate}
\alias{qm_validate}
\title{Validate input vector}
\usage{
qm_validate(.data, key, value)
}
\arguments{
\item{.data}{An sf object}

\item{key}{Quoted name of id variable to match input values to}

\item{value}{A vector of input values}
}
\value{
A logical scalar that is \code{TRUE} is all input values match values in the key variable.
}
\description{
This function ensures that the input vector values match valid values in a source shapefile.
}
